Fixed many issues showed by Coverity Scan

+++ mandelbulber2/src/opencl_engine_render_dof_phase2.cpp
@@ -237,7 +237,8 @@ bool cOpenClEngineRenderDOFPhase2::Render(
 		// requires initialization for all opencl devices
 		// requires optimalJob for all opencl devices
 
-		for (qint64 pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
+		for (qint64 pixelIndex = 0; pixelIndex < qint64(width) * qint64(height);
+				 pixelIndex += optimalJob.stepSize)
 		{
 			size_t pixelsLeft = width * height - pixelIndex;
