+++ mandelbulber2/src/opencl_engine_render_ssao.cpp
@@ -232,7 +232,8 @@ bool cOpenClEngineRenderSSAO::Render(cImage *image, bool *stopRequest)
 		// insert device for loop here
 		// requires initialization for all opencl devices
 		// requires optimalJob for all opencl devices
-		for (qint64 pixelIndex = 0; pixelIndex < width * height; pixelIndex += optimalJob.stepSize)
+		for (qint64 pixelIndex = 0; pixelIndex < qint64(width) * qint64(height);
+				 pixelIndex += optimalJob.stepSize)
 		{
 			size_t pixelsLeft = width * height - pixelIndex;
 			UpdateOptimalJobStart(pixelsLeft);
